Portafolios/Portafolio Implementación/NLP Module Project/requirements.txt
transformers
torch
flair
python-dotenv
pandas
matplotlib
requests
nltk